requirements.txt
django==3.2
djangorestframework
pyyaml
socketIO-client
requests
django-cors-headers
mysqlclient==2.1.1
PyJWT==1.7.1
daphne
channels
gunicorn
psycopg2
uuid